requirements.txt
Django==1.11.6
lorem==0.1.1
django-bootstrap-form==3.3
djangorestframework==3.7.1